abstract slice browser [skip ci]
fix mr vs ct handling [skip ci]

wip

revert qt abstraction changes, finished with slice browser abstraction

integrate with ieeg_locate, add tests

fix flake, fix tests?

fix codespell

test fix again

fix style

fix test

one more try

fix skip tests

another try for skiptest

try again with pyvistaqt in skip

oh, I think I fixed it, just missing a conftest arg

use fixture, hopefully fixes issues
